reorganised community tab related screens
