+++ lib/router.dart
@@ -10,8 +10,8 @@ import 'package:aura/view/community/meetup_listview.dart';
 import 'package:aura/view/community/thread_list_view.dart';
 import 'package:aura/view/settings/change_home_address_screen.dart';
 import 'package:aura/view/signin/signin_screen.dart';
-import 'package:aura/view/tabs/community/create_thread_view.dart';
-import 'package:aura/view/tabs/community/notifications_view.dart';
+import 'package:aura/view/community/create_thread_view.dart';
+import 'package:aura/view/community/notifications_view.dart';
 import 'package:aura/view/tabs/main_tab_bar.dart';
 import 'package:aura/view/settings/setting_screen.dart';
 import 'package:flutter/material.dart';
