+++ lib/view/community/create_thread_view.dart
