+++ lib/view/community/notifications_view.dart
